backend/requirements.txt
## Django packages
Django==5.2.2
PyYAML==6.0.2
uritemplate==4.1.1
Markdown==3.8
Pygments==2.19.1
django-filter==25.1
django-jazzmin==3.0.1
django-simple-history==3.8.0

# REST framework packages - COMMENTED OUT FOR DRF REMOVAL
djangorestframework==3.16
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
django-guardian==3.0.0

# Auth packages
django-allauth==65.8.1
python-decouple==3.8

# Testing packages
coverage==7.8.0
factory-boy==3.3.0

# Security packages
ggshield==1.43.0

# Network packages
gunicorn==23.0.0
uvicorn==0.35.0
psycopg2-binary==2.9.10
requests==2.32.4

# S3/MinIO Storage
django-storages[s3]==1.14.2
boto3==1.34.0

# Data processing
pandas==2.2.3
openpyxl==3.1.5

# Name similarity checking
rapidfuzz==3.6.1
jellyfish==1.0.3
num2words==0.5.13

# Dependency upgrades
# sqlparse>=0.5.0 # not directly required, pinned by Snyk to avoid a vulnerability
# zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability
# urllib3>=2.2.2 # not directly required, pinned by Snyk to avoid a vulnerability
